# api/docs/swagger.yaml
definitions:
...
  models.ChangePassword:
    properties:
      login:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  models.CreateCustomer:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      orders_count:
        type: integer
      phone:
        type: string
      unique_cars_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.GetAllCustomersResponse:
    properties:
      count:
        type: integer
      customers:
        items:
          $ref: '#/definitions/models.Customer'
        type: array
    type: object
...
  models.GetCustomerCars:
    properties:
      car_name:
        type: string
      duration:
        type: number
      order_created_at:
        type: string
      price:
        type: number
    type: object
  models.GetCustomerCarsResponse:
    properties:
      count:
        type: integer
      customer_cars:
        items:
          $ref: '#/definitions/models.GetCustomerCars'
        type: array
    type: object
...
  models.LoginCustomer:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      from_date:
        type: string
      id:
        type: string
      payment_status:
        type: boolean
      status:
        type: string
      to_date:
        type: string
      updated_at:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateCustomer:
    properties:
      address:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /customer:
    get:
      consumes:
      - application/json
      description: Retrieves information about all customers.
      parameters:
      - description: customers
        in: query
        name: search
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllCustomersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all customers
      tags:
      - customer
    post:
      consumes:
      - application/json
      description: Create a new customer
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CreateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a customer
      tags:
      - customer
  /customer/:
    patch:
      consumes:
      - application/json
      description: This api changes customer password by its login and password and
        returns message
      parameters:
      - description: Change Customer Password
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: customer change password
      tags:
      - customer
  /customer/{id}:
    delete:
      consumes:
      - application/json
      description: This api deletes a customer by its id and returns error or nil
      parameters:
      - description: customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: delete a customer by its id
      tags:
      - customer
    get:
      consumes:
      - application/json
      description: This api gets a customer by its id and returns its info
      parameters:
      - description: customer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: get a customer by its id
      tags:
      - customer
    put:
      consumes:
      - application/json
      description: This api updates a customer by its id and returns its id
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: update a customer
      tags:
      - customer
  /customer/cars:
    get:
      consumes:
      - application/json
      description: This api gets customer cars and returns their info
      parameters:
      - description: Customer ID
        in: query
        name: customerID
        required: true
        type: string
      - description: Car Name
        in: query
        name: carName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetCustomerCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: get customer's cars
      tags:
      - customer
  /customer/login:
    post:
      consumes:
      - application/json
      description: This api logs in customer account and returns message
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.LoginCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: customer login
      tags:
      - customer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"